Move out utility functions for mock data to utils
